calculate-veg-index-main/requirements.txt
sentinelsat
rasterio
shapely
python-dotenv
numpy
pyproj